+++ x/bank/types/codec.go
@@ -12,8 +12,8 @@ import (
 // RegisterLegacyAminoCodec registers the necessary x/bank interfaces and concrete types
 // on the provided LegacyAmino codec. These types are used for Amino JSON serialization.
 func RegisterLegacyAminoCodec(cdc *codec.LegacyAmino) {
-	cdc.RegisterConcrete(&MsgSend{}, "lbm-sdk/MsgSend", nil)
-	cdc.RegisterConcrete(&MsgMultiSend{}, "lbm-sdk/MsgMultiSend", nil)
+	cdc.RegisterConcrete(&MsgSend{}, "cosmos-sdk/MsgSend", nil)
+	cdc.RegisterConcrete(&MsgMultiSend{}, "cosmos-sdk/MsgMultiSend", nil)
 }
 
 func RegisterInterfaces(registry types.InterfaceRegistry) {