+++ x/slashing/types/msg_test.go
@@ -14,7 +14,7 @@ func TestMsgUnjailGetSignBytes(t *testing.T) {
 	bytes := msg.GetSignBytes()
 	require.Equal(
 		t,
-		`{"type":"lbm-sdk/MsgUnjail","value":{"address":"linkvaloper1v93xxeqn4h65f"}}`,
+		`{"type":"cosmos-sdk/MsgUnjail","value":{"address":"linkvaloper1v93xxeqn4h65f"}}`,
 		string(bytes),
 	)
 }